requirements.txt
pytelegrambotapi
pysocks